fix: unify settings layout prep when entering settings from outside
Clicking "Compose" in the emails tab without a connected account redirects
to the New Account settings page, but the navigation drawer was left in its
previous (collapsed / "main") state, producing a half-broken transition.

Three call sites were each duplicating the same "enter settings" prep
inline (memorize previous URL + drawer state, expand drawer, switch mobile
drawer to "settings"), and any other entry point that just called
useNavigateSettings — ComposeEmailButton, ComposeEmailCommand,
AIChatCreditsExhaustedMessage, several workflow/role components — skipped
it entirely and reproduced the bug.

Move the full prep into useOpenSettingsMenu (no-op when already on a
settings page so internal navigation doesn't clobber the memorized return
target), have useNavigateSettings delegate to it, and collapse the
duplicated logic in NavigationDrawerOtherSection and the multi-workspace
dropdown to a single call.
