+++ packages/twenty-front/src/hooks/useNavigateSettings.ts
@@ -1,9 +1,11 @@
+import { useOpenSettingsMenu } from '@/navigation/hooks/useOpenSettings';
 import { useNavigate } from 'react-router-dom';
 import { type SettingsPath } from 'twenty-shared/types';
 import { getSettingsPath } from 'twenty-shared/utils';
 
 export const useNavigateSettings = () => {
   const navigate = useNavigate();
+  const { openSettingsMenu } = useOpenSettingsMenu();
 
   return <T extends SettingsPath>(
     to: T,
@@ -14,6 +16,8 @@ export const useNavigateSettings = () => {
       state?: any;
     },
   ) => {
+    openSettingsMenu();
+
     const path = getSettingsPath(to, params, queryParams);
     return navigate(path, options);
   };
