+++ packages/twenty-front/src/modules/navigation/hooks/useOpenSettings.ts
@@ -1,17 +1,40 @@
+import { useIsSettingsPage } from '@/navigation/hooks/useIsSettingsPage';
 import { currentMobileNavigationDrawerState } from '@/navigation/states/currentMobileNavigationDrawerState';
 import { isNavigationDrawerExpandedState } from '@/ui/navigation/states/isNavigationDrawerExpanded';
+import { navigationDrawerExpandedMemorizedState } from '@/ui/navigation/states/navigationDrawerExpandedMemorizedState';
+import { navigationMemorizedUrlState } from '@/ui/navigation/states/navigationMemorizedUrlState';
+import { useAtomState } from '@/ui/utilities/state/jotai/hooks/useAtomState';
 import { useSetAtomState } from '@/ui/utilities/state/jotai/hooks/useSetAtomState';
+import { useLocation } from 'react-router-dom';
 
+// Prepares the layout when entering the settings area from outside.
+// Memorizes the previous URL and drawer state so "Exit Settings" can restore
+// them, expands the desktop drawer, and switches the mobile drawer to
+// "settings". A no-op when already on a settings page so internal navigation
+// doesn't clobber the memorized return target.
 export const useOpenSettingsMenu = () => {
-  const setIsNavigationDrawerExpanded = useSetAtomState(
-    isNavigationDrawerExpandedState,
+  const location = useLocation();
+  const isSettingsPage = useIsSettingsPage();
+  const [isNavigationDrawerExpanded, setIsNavigationDrawerExpanded] =
+    useAtomState(isNavigationDrawerExpandedState);
+  const setNavigationDrawerExpandedMemorized = useSetAtomState(
+    navigationDrawerExpandedMemorizedState,
+  );
+  const setNavigationMemorizedUrl = useSetAtomState(
+    navigationMemorizedUrlState,
   );
   const setCurrentMobileNavigationDrawer = useSetAtomState(
     currentMobileNavigationDrawerState,
   );
 
   const openSettingsMenu = () => {
+    if (isSettingsPage) {
+      return;
+    }
+
+    setNavigationDrawerExpandedMemorized(isNavigationDrawerExpanded);
     setIsNavigationDrawerExpanded(true);
+    setNavigationMemorizedUrl(location.pathname + location.search);
     setCurrentMobileNavigationDrawer('settings');
   };
